requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0

# AI/ML Frameworks
torch==2.1.1
torchvision==0.16.1
diffusers==0.24.0
transformers==4.36.2
accelerate==0.25.0
safetensors==0.4.1
xformers==0.0.23

# Image Generation
Pillow==10.1.0
opencv-python==4.8.1.78
rembg==2.0.50
controlnet-aux==0.0.7
mediapipe==0.10.8

# Video Generation
imageio==2.33.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
decord==0.6.0

# Text/Novel Generation
openai==1.3.7
langchain==0.0.352
langchain-community==0.0.7
tiktoken==0.5.2

# Human Body Mapping
smplx==0.1.28
chumpy==0.70
trimesh==4.0.5
pyrender==0.1.45

# Income Stream Automation
stripe==7.8.0
celery==5.3.4
redis==5.0.1
schedule==1.2.1
pandas==2.1.4
numpy==1.26.2

# Web UI
gradio==4.11.0
jinja2==3.1.2
python-multipart==0.0.6

# Utilities
aiohttp==3.9.1
httpx==0.25.2
tenacity==8.2.3
loguru==0.7.2
tqdm==4.66.1
pyyaml==6.0.1
rich==13.7.0
